Update lge blobs to german V20c 07-10-2014
LG used a custom libjpeg in stock rom and the camera blob is linked with that.
It cannot encode pictures with the source-built one. So the camera blob is hexedited
to use lgejpeg.so instead of libjpeg.so. So you should pull libjpeg.so from V20C as well,
put it on device as lgejpeg.so and modify the camera blob to use the renamed one.